core/requirements.txt
aniso8601==7.0.0
asgiref==3.2.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
Django==3.0.4
django-cors-headers==3.2.1
django-fernet-fields==0.6
django-graphql-jwt==0.3.0
django-jsonfield==1.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
geoip2==3.0.0
gql==0.4.0
graphene==2.1.8
graphene-django==2.9.0
graphql-core==2.3.1
graphql-relay==2.0.1
idna==2.9
jsonfield==3.1.0
maxminddb==1.5.2
numpy==1.18.2
pandas==1.0.3
promise==2.3
psycopg2-binary==2.8.4
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
stripe==2.45.0
urllib3==1.25.8
